salesforce/requirements.txt
asgiref==3.5.1
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-guardian==2.4.0
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
lxml==4.8.0
ndg-httpsclient==0.5.1
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
salesforce==0.0.1
salesforce-api==0.1.45
simple-salesforce==1.11.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
url-normalize==1.4.3
urllib3==1.26.9
xmltodict==0.13.0
zeep==4.1.0
